playerのコンボ攻撃が動くように変更。アニメーションコントローラーでnotifyAnimationIndexの更新タイミングを変更。

+++ Source/Game/Actors/Player/TestPlayer.cpp
@@ -107,7 +107,7 @@ void TestPlayer::Initialize(const Transform& transform)
 void TestPlayer::Update(float elapsedTime)
 {
     Character::Update(elapsedTime);
-	clip_index = GetBodyAnimationController()->GetAnimationClip();
+	clip_index = static_cast<int>(GetBodyAnimationController()->GetAnimationClip());
 	std::string currentAnimationName = GetBodyAnimationController()->GetCurrentAnimationName();
 #if 0
 	bool pressing_shift = (GetKeyState(VK_SHIFT) & 0x8000);
